# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Food Log Backend API
  description: |
    REST API for the Food Log application. The service accepts meal photo uploads,
    predicts nutrition information, and persists meal history in DynamoDB or SQLite.
  version: 1.0.0
servers:
  - url: https://api.foodlog.example.com
    description: Production
  - url: http://localhost:5000
    description: Local development
tags:
  - name: Authentication
  - name: Predictions
  - name: History
  - name: Health
paths:
  /predict:
    post:
      tags: [Predictions]
      summary: Upload a meal photo and receive nutrition predictions
      operationId: uploadMealPhoto
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photo:
                  type: string
                  format: binary
                  description: Image file of the meal. The field may also be named `image`.
              required:
                - photo
      responses:
        "200":
          description: Prediction generated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PredictionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
  /auth/signup:
    post:
      tags: [Authentication]
      summary: Register a new user
      operationId: signUp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      tags: [Authentication]
      summary: Authenticate an existing user
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: Login succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      tags: [Authentication]
      summary: Fetch details for the authenticated user
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User details returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /history:
    get:
      tags: [History]
      summary: List stored meal predictions
      operationId: listHistory
      responses:
        "200":
          description: History retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/StoredMeal"
  /health:
    get:
      tags: [Health]
      summary: Health-check endpoint
      operationId: healthCheck
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  timestamp:
                    type: string
                    format: date-time
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: S3curePass!
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT for subsequent authenticated requests.
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
    PredictionResponse:
      type: object
      properties:
        image_url:
          type: string
          format: uri
        food:
          type: string
        calories:
          type: integer
          example: 420
        ingredients:
          type: array
          items:
            type: string
        nutrition_facts:
          $ref: "#/components/schemas/NutritionFacts"
        metadata:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
        inference_source:
          type: string
          description: Indicates whether the ML service or local fallback produced the result.
        ml_service_error:
          type: string
          description: Populated when the remote ML service call fails.
    StoredMeal:
      allOf:
        - $ref: "#/components/schemas/PredictionResponse"
        - type: object
          properties:
            id:
              type: string
              description: Unique identifier for the stored record.
    NutritionFacts:
      type: object
      properties:
        calories:
          type: integer
        carbohydrates:
          type: integer
        proteins:
          type: integer
        fats:
          type: integer
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
          nullable: true
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServerError:
      description: Unexpected error occurred
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"